IQA-111: Move api-routing-configuration to api-rootfolder and cleanup app-context
