+++ web-app/src/__init__.py
@@ -30,10 +30,9 @@ def create_app():
     with app.app_context():
         from .services.UserService import UserService
 
-        from .template_extensions import error_handlers
-        from .template_extensions import context_preprocessors
+        from .template_extensions import error_handlers, context_preprocessors
 
-        from .api.v1 import api_routing_configuration
+        from .api import api_routing_configuration
 
         from .modules import routing_configuration
