+++ web-app/src/api/api_routing_configuration.py
@@ -1,8 +1,6 @@
 from flask import current_app as app
 
 
-from . import admin
+from .v1 import admin, subjects
 app.register_blueprint(admin.api_v1__admin_controller)
-
-from . import subjects
 app.register_blueprint(subjects.api_v1__subjects_controller)
